make errors for cards

+++ controllers/cards.js
@@ -4,39 +4,54 @@ const {
   HTTP_STATUS_OK, HTTP_STATUS_NOT_FOUND,
   HTTP_STATUS_CREATED,
 } = require('http2').constants;
-const Card = require('../models/card');
 
-module.exports.createCard = (req, res) => {
+// const ConflictError = require('../errors/ConflictError');
+const NotFoundError = require('../errors/NotFoundError');
+// const UnauthorizedError = require('../errors/UnauthorizedError');
+const BadRequestError = require('../errors/BadRequestError');
+
+const Card = require('../models/card');
+const ConflictError = require('../errors/ConflictError');
+
+module.exports.createCard = (req, res, next) => {
   const { name, link } = req.body;
 
   return Card.create({ name, link, owner: req.user._id })
     .then((r) => res.status(HTTP_STATUS_CREATED).send(r))
     .catch((e) => {
       if (e.name === 'ValidationError') {
-        return res.status(HTTP_STATUS_BAD_REQUEST).send({ message: 'Переданы некорректные данные при создании карточки.' });
+        next(new BadRequestError('Переданы некорректные данные при создании карточки.'));
+        return;
       }
-      return res.status(HTTP_STATUS_INTERNAL_SERVER_ERROR).send({ message: 'Непредвиденная ошибка на сервере.' });
+      next(e);
     });
 };
 
-module.exports.getCards = (req, res) => Card.find({})
+module.exports.getCards = (req, res, next) => Card.find({})
   .then((r) => res.status(HTTP_STATUS_OK).send(r))
-  .catch(() => res.status(HTTP_STATUS_INTERNAL_SERVER_ERROR).send({ message: 'Непредвиденная ошибка на сервере.' }));
+  .catch(next);
 
-module.exports.deleteCardById = (req, res) => {
+module.exports.deleteCardById = (req, res, next) => {
   const { cardId } = req.params;
-  return Card.findByIdAndRemove(cardId)
+  return Card.findById(cardId).orFail(new NotFoundError('Такой карточки не существует!'))
     .then((r) => {
       if (r === null) {
-        return res.status(HTTP_STATUS_NOT_FOUND).send({ message: 'Карточка не найдена!' });
+        throw new NotFoundError('Карточка не найдена!');
+        // return res.status(HTTP_STATUS_NOT_FOUND).send({ message: 'Карточка не найдена!' });
+        // дублируется
       }
-      return res.status(HTTP_STATUS_OK).send(r);
+      if (`${r.owner}` !== req.user._id) {
+        throw new ConflictError('Сожалеем, но Вы можете удалять только свои карточки!');
+      }
+      return Card.findByIdAndRemove(cardId);
+      // return res.status(HTTP_STATUS_OK).send(r);
     })
     .catch((e) => {
       if (e.name === 'CastError') {
-        return res.status(HTTP_STATUS_BAD_REQUEST).send({ message: 'Такой карточки не существует!' });
+        next(new BadRequestError('Такой карточки не существует!'));
+        return;
       }
-      return res.status(HTTP_STATUS_INTERNAL_SERVER_ERROR).send({ message: 'Непредвиденная ошибка на сервере.' });
+      next(e);
     });
 };
